refactor: UIImage resize extensions 파일로 이동
